This code deals with issue 19: https://github.com/bitpay/magento2-plugin/issues/19
Config's Resouremodel's saveconfig function was called with an incorrect scope type leading to an application crash after saving the value to the core config database.

/Model/MagentoStorage and /Helper/Data now have the scope config set to SCOPE_TYPE_DEFAULT.

The goal of the previous code is unclear as the storemanager was called to request the current store; which is unrelated to the chosen scope in the admin for setting configuration values.